Refactor models and moved session helper functions to a new helper directory
